feat: 🌟 add @radix-ui/react-radio-group dependency and update AddEventDialog and EditEventDialog to use ColorPicker component
